Additoinal proc macro
