+++ rml-macro/src/main.rs
@@ -1,12 +1,5 @@
-use rml_macro::TensorCreator;
+use rml_macro::{TensorCreator, TensorDebug};
 
-#[derive(TensorCreator)]
-enum Example {
-    Col1(f32),
-    Col2(i32),
-    Col3(String)
-
-}
 
 fn main() {
 }
